requirements.txt
asgiref==3.8.1
Django==5.0.4
django-allauth==0.62.1
django-crispy-forms==2.1
geographiclib==2.0
geopy==2.4.1
pillow==10.3.0
sqlparse==0.4.4
